requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
Django==4.2.6
django-otp==1.2.4
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
magic-filter==1.0.12
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
redis==5.0.1
sqlparse==0.4.4
types-PyMySQL==1.1.0.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.7
typing_extensions==4.7.1
yarl==1.9.2
